requirements_reference.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_ipyw_jlab_nb_ext_conf=0.1.0=py37_0
anaconda-client=1.7.2=py_0
anaconda-navigator=1.9.2=py37_0
argon2-cffi=20.1.0=py37h4ab8f01_2
asciitree=0.3.3=py_2
async_generator=1.10=py_0
attrs=20.2.0=pyh9f0ad1d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
beautifulsoup4=4.9.3=pyhb0f4dca_0
bids-validator=1.7.2=pypi_0
bioe300b=0.1=pypi_0
bleach=3.2.1=pyh9f0ad1d_0
bottleneck=1.3.2=py37h3a3b6f7_2
brotlipy=0.7.0=py37h0013d47_1001
ca-certificates=2020.12.5=h5b45459_0
cairo=1.16.0=hd28d34b_1006
certifi=2020.12.5=py37h03978a9_1
cffi=1.14.3=py37hd6b71e5_1
chardet=3.0.4=py37hf50a25e_1008
click=8.0.1=pypi_0
clyent=1.2.2=py_1
colorama=0.4.4=pyh9f0ad1d_0
conda=4.9.2=py37h03978a9_0
conda-build=3.15.1=py37_0
conda-env=2.6.0=1
conda-package-handling=1.7.2=py37h6dbccfb_0
console_shortcut=0.1.1=3
cryptography=3.2.1=py37hd8e9650_0
cycler=0.10.0=py_2
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
dill=0.3.4=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.3=py37hc8dfbb8_1002
filelock=3.0.12=pyh9f0ad1d_0
freetype=2.10.4=h546665d_0
gettext=0.19.8.1=hfbb10ce_1004
glob2=0.7=py_0
icu=67.1=h33f27b4_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipdb=0.13.9=pypi_0
ipykernel=5.3.4=py37h7b7c402_1
ipython=7.19.0=py37heaed05f_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.5.1=pyh9f0ad1d_1
jedi=0.17.2=py37hc8dfbb8_1
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.17.0=py_0
jpeg=9d=h8ffe710_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_contrib_core=0.3.3=py_2
jupyter_contrib_nbextensions=0.5.1=py37hc8dfbb8_1
jupyter_core=4.6.3=py37hc8dfbb8_2
jupyter_highlight_selected_word=0.2.0=py37h03978a9_1002
jupyter_latex_envs=1.4.6=py37hc8dfbb8_1001
jupyter_nbextensions_configurator=0.4.1=py37h03978a9_2
jupyterlab=2.2.9=py_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=1.2.0=py_0
kiwisolver=1.3.1=py37h8c56517_0
libblas=3.8.0=20_mkl
libcblas=3.8.0=20_mkl
libclang=10.0.1=default_hf44288c_1
libffi=3.2.1=ha925a31_1007
libflang=5.0.0=h6538335_20180525
libglib=2.66.2=h35efcdc_0
libiconv=1.16=he774522_0
liblapack=3.8.0=20_mkl
libpng=1.6.37=h1d00b33_2
libsodium=1.0.18=h62dcd97_1
libtiff=4.1.0=hc10be44_6
libxml2=2.9.10=hf5bbc77_3
libxslt=1.1.33=h65864e5_2
llvm-meta=5.0.0=0
lxml=4.6.2=py37hd07aab1_1
lz4-c=1.9.2=h62dcd97_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py37h0013d47_2
matplotlib=3.3.2=py37h03978a9_1
matplotlib-base=3.3.2=py37h3379fd5_1
menuinst=1.4.16=py37hc8dfbb8_1
mistune=0.8.4=py37h4ab8f01_1002
mkl=2020.2=256
msys2-conda-epoch=20160418=1
navigator-updater=0.2.1=py37_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37hc8dfbb8_2
nbformat=5.0.8=py_0
nest-asyncio=1.4.2=pyhd8ed1ab_0
networkx=2.5=py_0
neuprint-python=0.4.11=py_0
neurom=1.5.3.dev1=dev_0
nibabel=3.2.1=pypi_0
nilearn=0.8.0=pypi_0
notebook=6.1.4=py37hc8dfbb8_1
num2words=0.5.10=pypi_0
numpy=1.19.4=py37hd20adf4_1
olefile=0.46=pyh9f0ad1d_1
openmp=5.0.0=vc14_1
openssl=1.1.1j=h8ffe710_0
packaging=20.4=pyh9f0ad1d_0
pandas=0.23.4=pypi_0
pandoc=2.11.0.4=hf4a77e7_0
pandocfilters=1.4.2=py_1
parso=0.7.1=pyh9f0ad1d_0
patsy=0.5.1=py_0
pcre=8.44=ha925a31_0
pickleshare=0.7.5=py_1003
pillow=8.0.1=py37hdec93a9_0
pip=20.2.4=pypi_0
pixman=0.40.0=h8ffe710_0
pkginfo=1.6.1=pyh9f0ad1d_0
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.8=py_0
psutil=5.7.3=py37h0013d47_0
pybids=0.13.1=pypi_0
pycairo=1.20.0=py37hd14c632_1
pycosat=0.6.3=py37h0013d47_1005
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.2=py_0
pymef=1.2.1=pypi_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py37h1834ac0_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.17.3=py37h4ab8f01_1
pysocks=1.7.1=py37hf50a25e_2
python=3.7.0=hea74fb7_0
python-dateutil=2.8.1=py_0
python-igraph=0.8.2=py37hc69c63d_0
python_abi=3.7=1_cp37m
pytz=2020.4=pyhd8ed1ab_0
pywin32=228=py37h4ab8f01_0
pywinpty=0.5.7=py37hc8dfbb8_1
pyyaml=5.3.1=py37h0013d47_1
pyzmq=19.0.2=py37h453f00a_2
qt=5.12.9=hb2cf2c5_0
qtpy=1.9.0=py_0
requests=2.24.0=pyh9f0ad1d_0
ruamel_yaml=0.15.80=py37h0013d47_1003
scikit-learn=0.23.2=py37heb15398_2
scipy=1.5.3=py37h6db1a17_0
seaborn=0.11.0=h57928b3_1
seaborn-base=0.11.0=pyhd8ed1ab_1
send2trash=1.5.0=py_0
setuptools=49.6.0=py37hf50a25e_2
six=1.15.0=pyh9f0ad1d_0
soupsieve=2.0.1=py_1
sqlalchemy=1.3.24=pypi_0
sqlite=3.33.0=he774522_1
statsmodels=0.12.1=py37hda49f71_1
terminado=0.9.1=py37hc8dfbb8_1
testpath=0.4.4=py_0
texttable=1.6.3=pyh9f0ad1d_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_1
toml=0.10.2=pypi_0
tornado=6.1=py37hcc03f2d_0
tqdm=4.51.0=pyh9f0ad1d_0
traitlets=5.0.5=py_0
ujson=4.0.1=py37h1834ac0_1
urllib3=1.25.11=py_0
vc=14.1=h869be7e_1
vs2015_runtime=14.16.27012=h30e32a0_2
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.35.1=pyh9f0ad1d_0
widgetsnbextension=3.5.1=py37hc8dfbb8_4
win_inet_pton=1.1.0=py37hc8dfbb8_1
wincertstore=0.2=py37hc8dfbb8_1005
winpty=0.4.3=4
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
zeromq=4.3.2=ha925a31_4
zipp=3.4.0=py_0
zlib=1.2.11=h62dcd97_1010
zstd=1.4.5=h1f3a1b7_2
